fix lint

+++ fairseq2n/src/fairseq2n/data/image/png_decoder.cc
@@ -130,7 +130,7 @@ png_decoder::operator()(data &&d) const
         for (png_uint_32 i = 0; i < height; ++i) {
             png_read_row(png_ptr, image_data, nullptr);
             for (size_t j = 0; j < rowbytes; ++j) {
-                image_data[j] = (int32_t)image_data[j];
+                image_data[j] = static_cast<float>(image_data[j]);
             }
             image_data += rowbytes;
         }
